refactor: move tracing to shared
