+++ storagemarket/provider.go
@@ -23,8 +23,8 @@ import (
 	"github.com/filecoin-project/boost/storagemarket/types"
 	smtypes "github.com/filecoin-project/boost/storagemarket/types"
 	"github.com/filecoin-project/boost/storagemarket/types/dealcheckpoints"
-	"github.com/filecoin-project/boost/tracing"
 	"github.com/filecoin-project/boost/transport"
+	"github.com/filecoin-project/boostd-data/shared/tracing"
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-fil-markets/piecestore"
 	"github.com/filecoin-project/go-fil-markets/shared"
